Add RemoveCart endpoint to CartAPIController
Added a new HTTP POST endpoint `RemoveCart` to the `CartAPIController` class in the `Mango.Services.ShoppingCartAPI.Controllers` namespace. This endpoint allows for the removal of a cart item by its `cartDetailsId`. If the cart item being removed is the last item in the cart, the corresponding cart header is also removed. The method includes exception handling and updates the response object accordingly.